feat: svg 컴포넌트 작성

+++ src/app/(not-logged-in)/(oauth)/authenticate/CheckAuth.ts
@@ -14,7 +14,7 @@ export default function CheckAuth() {
   useEffect(() => {
     const isLoggedin = status === "authenticated" && sidCookie !== undefined;
 
-    if (!isLoggedin && pathname !== "/oauth") {
+    if (status !== "loading" && !isLoggedin && pathname !== "/oauth") {
       deleteCookie("connect.sid");
       deleteCookie("loggedIn");
       router.push("/oauth");
